fix some responsive text styles (#227)
This PR moves the ReactiveElement to components and improves type checking, then uses it across all pages where needed. Also makes a couple minor style tweaks to match mocks after talking with Tahnee.
